fix websocket pong response, code refactor

+++ gps/location.go
@@ -1,6 +1,6 @@
 package gps
 
 type Location struct {
-	Latitude  float32 `json:"latitude"`
-	Longitude float32 `json:"longitude"`
+	Latitude  float64 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
 }
